// src/interfaces/PokemonType.d.ts
export type PokemonType = (
    "normal" |
    "fire" |
    "fighting" |
    "water" |
    "flying" |
    "grass" |
    "poison" |
    "electric" |
    "ground" |
    "psychic" |
    "rock" |
    "ice" |
    "bug" |
    "dragon" |
    "ghost" |
    "dark" |
    "steel" |
    "fairy" |
    "stellar" |
    "???"
);